[Rust][Tools] Stub Service + Schema Registry Stub (#690)
# Context

The schema registry service deployed by AIO has external dependencies,
making it challenging to use for inner loop testing. To address this,
the stub service introduced in this PR provides a simple, local
implementation of the schema registry service to accelerate inner loop
development.

# Changes

- Adds a new crate in the `tools` folder for the stub service.
- Introduces a simple CI job to build the stub service.
- Stub Service:
- Includes multiple stubs for different services in the `Services` crate
that have external dependencies and are difficult to simulate.
  - Each service has its own session to approximate real-world behavior.
- Each service logs its state and activity to an output directory
specified by an environment variable.
- The output format is `[Environment
variable]/stub_service_[Timestamp]/[Service Name]/[State / Log
name].json`.
- Schema Registry Stub:
- Implements a basic schema registry service supporting `Get` and `Put`
operations.
- Maintains an internal hashmap where the key is a schema's hash, and
the value is a set of schema versions.
- Outputs logs to the `SchemaRegistry` folder within the current
instance of the output directory.
- Supports only happy path interactions (valid `Get` and `Put`
operations).

# Future Improvements

- Add proper error handling for schema registry service errors. Plans
are underway to define these errors in the schema registry's DTDL, so
error handling is not implemented in this version.
- Integrate the stub service with the SDK testing pipeline.
- Add a way to maintain dtdls in sync with services.

# Comments

A version of the schema registry without external dependencies is under
evaluation. If implemented, this stub service will be deprecated and
removed.