[Rust][Services] Split leased_lock client (ADR 20) and add lease & lock auto-renewal (#759)
According to ADR 20 this leased lock client should be split into two
layers, lease and lock clients. The lease client will have the bulk of
logic for acquiring, renewing and releasing a key. The lock layer will
present the expected semantics for locking operations.
Additionally auto-renewal of leases was not yet implemented in the
leased lock client. included in this PR is the logic to perform the
auto-renewal of leases/locks.

---------

Co-authored-by: Ryan Winter <ryanwinter@outlook.com>