[Rust] [Connector] Get assets implementation (#789)
This PR:
- Adds implementation for getting Assets and Datasets and reporting
their statuses and message schemas using the base connector
- updates sample to test this flow **NOTE: I still don't think this
sample will stick around forever, it's just a way to test this as we
develop**
- adds derive-getters for base connector structs to ensure that
consuming logic can't modify these fields
- adds schema registry integration to the base connector to support the
message schema reporting flow
- removes the data transformer and related generics from the base
connector
- switches to using refs for all names (device, endpoint, asset, and
dataset) to avoid having a bunch of name fields on every struct and to
keep them symmetrical
- updates some data sharing strategies for device and asset fields

This PR doesn't (will come in future PRs):
- [p0] block forwarding data if the MessageSchema has not been added
yet.
- have correct data sharing strategies for specification fields since
updates aren't supported yet. These will look more like status fields in
the future
- share the device status down through the dataset level, it should in
the future for completion
- have similar create observation flows for datasets as there are for
devices and assets. This may come in the future as we determine whether
this is helpful or adds bloac
- have a macro for creating new struct definitions that are similar to
ADR struct definitions
- have support for events, streams, and management groups in addition to
datasets. This change will likely group these implementations under an
enum
- track the MessageSchema and MessageSchemaReference (fully) within the
dataset. This should be added in the future so it's easier to determine
whether a new MessageSchema is the same as the current one or not
- have the DatasetRef, AssetRef, and DeviceEndpointRefs in the correct
places - currently they are defined in different locations and should be
consolidated to the adr client