[Rust][Connector] Add tokio rwlock to ensure all dataset status updates reach service (#925)
This PR changes the sync rwlock to a tokio rwlock and holds the lock for
longer when writing to protect against multiple dataset clients updating
the asset status concurrently. This ensures only one status can be
modified and reported at a time so that all status updates get sent to
the service. (the service only accepts a full status replace, they don't
do any patch handling of status updates). This may add some latency
around reporting statuses as they can't happen concurrently, but this is
better than losing status information. Changes in this PR are mostly
restructuring to allow for this mutex to be held in the correct places

Note that this same change does not need to be made on the device
client, because the device status can only be reported from the device
(unlike the asset status, which can be reported from multiple dataset
clients)