[Rust] Clean up Redundant Clones (#931)
Did a quick pass with [Clippy Lint Redundant
Clone](https://rust-lang.github.io/rust-clippy/master/#redundant_clone)
set to warn and fixed all of the correct warnings that came up. We can't
leave this on by default because it isn't super stable and detects some
needed clones, but can be used as a local tool to help with reducing
extra `clone()`s. Note that this also doesn't catch everything, so there
may be more that could be cleaned up, and this still needs to be
manually thought about, even if this warning is on