Refactor ProtocolCompiler into library and tool that uses library (#974)
This is purely a refactoring; there is no functional change to the
ProtocolCompiler tool. The purpose of this change is to enable use of
the ProtocolCompiler functionality from a service (in particular, from
the AI MCP server I am prototyping).

Does anyone have time for a quick once over? @timtay-microsoft,
@clecompt-msft, @vaavva?