[Rust] Support AIO peristence (#976)
This change adds support for the AIO persistence flag to the Rust SDK in
the following places: session (connect), telemetry send, and state-store
set. It does not add it to the generic MQTT publish, since the publish
options do not currently extend the underlying MQTT options.

Minor breaking changes:
* Added a field to `SetOptions`, requiring adding `default` where it was
not already in use.
* Made the fields of `SessionOptions` private, requiring the (existing)
builder to be used.
* `aio_broker_features` is now an optional struct (pass `None` instead
of `false` to disable).