[Rust][Connector][Services] ADR EventGroups Support (and other 2510 asset ADR updates) (#1023)
This PR:
- Updates the DTDL for the ADR service for changes coming for Assets for
2510. This (mainly) includes the addition of EventGroups, but there are
also a few other minor field adjustments
- Regenerates ADR service code off of the new DTDL changes (cleaned up a
few stale files as well)
- Updates the ADR Client Service SDK to handle the new DTDL
- Updates the Connector SDK to handle the new DTDL
- Updates how events are handled in a DataOperation - they now use an
EventSpecification struct that contains the Event definition and that
event's EventGroupSpecification. Also changes the "name" of events to be
a tuple to be able to dedupe Events on different EventGroups with the
same name
- Updates status reporting to include the eventGroup layer (although
eventGroups don't have their own status reporting)
- Updates the DataOperationRef and DataOperation to formats that better
handle eventGroups
- Updates base connector sample (most changes are some log redundancy
simplifications that are only sorta tied to the SDK changes)
 
This PR DOES NOT:
- support eventGroup.default_events_destinations. This field is not used
to inform the destination of an event currently
- support separate eventGroup notifications - any updates to an
eventGroup definition will be reported on each of its Events and cannot
be received separately. Note that this limitation does not negatively
affect this scenario: updates to event1 on eventGroupA do not cause a
notification on event2 on eventGroupA

State of testing:
- Service is not available yet, so none of this has been able to be
tested yet beyond compilation. Before this is released, these changes
must be tested with the service